refactor(act): promote lru-map out of internal/ to src/lru-map.ts
LruMap and LruSet aren't pipeline internals — they're a generic
bounded LRU utility. Two real consumers exist:

- adapters/in-memory-cache.ts (production code, was reaching into
  internal/lru-map.js with a deep import — the lone production-side
  bypass of the barrel)
- internal/correlate-cycle.ts (still uses LruSet for the subscribed-
  streams cap; now imports it from the new top-level location)

Move to libs/act/src/lru-map.ts alongside src/utils.ts, update both
import sites + the spec. After this, internal/ holds only orchestrator-
pipeline modules — adapters and tests no longer have a legitimate
reason to deep-import into it.
