+++ libs/act/src/internal/correlate-cycle.ts
@@ -14,6 +14,7 @@
  * @internal
  */
 
+import { LruSet } from "../lru-map.js";
 import { store } from "../ports.js";
 import type {
   Query,
@@ -23,7 +24,6 @@ import type {
   Schemas,
 } from "../types/index.js";
 import type { DrainOps } from "./drain.js";
-import { LruSet } from "./lru-map.js";
 
 /**
  * Static resolver target collected at build time. Subscribed once during
